Fix pyright standard mode type errors and add CI type checking
- Fix OpenCV Scalar type mismatches by using tuples instead of bare ints
  for color/borderValue args (cv2.fillPoly, cv2.line, cv2.remap, etc.)
- Fix numpy floating return type in _mad_sigma_1d
- Add py.typed marker for PEP 561 compliance
- Add pyright step to test workflow

+++ src/lensboy/calibration/calibrate.py
@@ -75,7 +75,7 @@ def _mad_sigma_1d(x: np.ndarray) -> float:
     med = np.median(x)
     mad = np.median(np.abs(x - med))
     # 1.4826 = 1 / Phi^{-1}(0.75)  (MAD->sigma for 1D normal)
-    return 1.4826 * mad
+    return float(1.4826 * mad)
 
 
 def _robust_sigma_xy(residuals: list[np.ndarray]) -> float:
